Add files via upload
If you download the LitXUMLClassDiagram_1.0.xml file in this download you should be able to run the draw.io version to add and remove stuff by going to https://www.draw.io/ clicking "Open existing diagram" and then selecting the XML file you just downloaded. This is 90% of the complete draft of the UML Class Diagram all that is missing is some aggregation lines as well as implementing the wow factor. I am gonna try to finish off these last parts tomorrow (hopefully before 4) and then spend some time editting it and making it even more readable (this should be done by 8 tomorrow at the latest). I hope that after that point that someone (or many someones) can thoroughly look over it and maybe we could discuss ways to improve my diagram or you can use this download to just change it yourself and then reupload it to GIT.
Richard Davidson (The Great)
P.S You can contact me via Hangouts or by texting/calling me at 204 963 0319 I have a minor surgery on Wednesday so I doubt ill be able to work on this wednesday but I should be good by thursday to discuss/work on the UML some more